[docker] Fix error removing temporal files from the image
Files were copied with COPY docker command which uses
'root' as user. However, the current user running commands
is 'opensearch', so the file won't be removed when it was
copied to '/tmp/' because it was created by 'root'.
This commit fixes the issue copying the file into a place
where 'opensearch' user has enough privileges to remove it.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>